some testing for new code

+++ data/indexed_dataset.py
@@ -11,7 +11,7 @@ import struct
 import numpy as np
 import torch
 
-from tokenizer import Tokenizer
+from .tokenizer import Tokenizer
 
 
 def read_longs(f, n):
@@ -30,7 +30,7 @@ dtypes = {
     3: np.int16,
     4: np.int32,
     5: np.int64,
-    6: np.float,
+    6: float,
     7: np.double,
 }
 
@@ -181,7 +181,7 @@ class IndexedDatasetBuilder(object):
         np.int16: 2,
         np.int32: 4,
         np.int64: 8,
-        np.float: 4,
+        float: 4,
         np.double: 8
     }
